{{cookiecutter.pypi_package_name}}/pyproject.toml
[project]
name = "{{cookiecutter.project_slug}}"
version = "{{ cookiecutter.first_version }}"
description = "{{cookiecutter.project_short_description}}"
readme = "README.md"
authors = [
  {name = "{{cookiecutter.full_name}}", email = "{{cookiecutter.email}}"}
]
maintainers = [
  {name = "{{cookiecutter.full_name}}", email = "{{cookiecutter.email}}"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
    "django",
    "celery",
    "djangorestframework",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "ruff",  # linting
]

[project.urls]
bugs = "https://github.com/{{cookiecutter.__gh_slug}}/issues"
changelog = "https://github.com/{{cookiecutter.__gh_slug}}/blob/master/changelog.md"
homepage = "https://github.com/{{cookiecutter.__gh_slug}}"


[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
